--- apps/web/src/app/ab-test-a/_components/ABStepIndicator.tsx
@@ -1,37 +1,0 @@
-'use client';
-
-import { cn } from '@/lib/utils';
-
-import { SURVEY_STEPS } from '../_constants';
-
-interface ABStepIndicatorProps {
-  currentStep: number;
-}
-
-/**
- * A/B 테스트 B안 스텝 인디케이터
- * 3개의 점으로 현재 진행 단계를 표시합니다.
- * 각 dot 클릭 시 해당 스텝으로 이동할 수 있습니다.
- */
-export default function ABStepIndicator({ currentStep }: ABStepIndicatorProps) {
-  return (
-    <div className="mb-6 flex gap-2">
-      {Array.from({ length: SURVEY_STEPS }, (_, index) => {
-        const stepNumber = index + 1;
-        const isActive = currentStep === stepNumber;
-
-        return (
-          <button
-            type="button"
-            key={stepNumber}
-            className={cn(
-              'h-2 w-2 cursor-pointer rounded-[50%] border-none bg-gray-200 p-0',
-              isActive && 'bg-gray-900'
-            )}
-            aria-label={`Step ${stepNumber} ${isActive ? '(현재)' : ''}`}
-          />
-        );
-      })}
-    </div>
-  );
-}
