--- apps/web/src/app/ab-test-a/_components/index.ts
@@ -1,4 +1,0 @@
-export { default as ABCardContainer } from './ABCardContainer';
-export { default as ABPageLayout } from './ABPageLayout';
-export { default as ABProgressBar } from './ABProgressBar';
-export { default as ABStepIndicator } from './ABStepIndicator';
