--- apps/web/src/app/ab-test-a/_components/steps/index.ts
@@ -1,8 +1,0 @@
-export { default as ABAllergyStep } from './aAllergyStep';
-export { default as ABIntroSlider } from './aIntroSlider';
-export { default as ABIntroStep } from './aIntroStep';
-export { default as ABuseIntroSlider } from './aUseIntroSlider';
-export { default as ConditionStep } from './aConditionStep';
-export { default as IngredientsStep } from './IngredientsStep';
-export { default as IntroStep } from './IntroStep';
-export { default as RecipeResultStep } from './RecipeResultStep';
