--- apps/web/src/app/ab-test-a/_hooks/index.ts
@@ -1,1 +1,0 @@
-export { useABTestStep } from './useABTestStep';
