--- apps/web/src/app/ab-test-a/_styles/ab-test.css
@@ -1,21 +1,0 @@
-/* 인트로 페이지네이션 스타일 - 밝은 배경용 (슬라이드 2,3) */
-.ab-intro-pagination .swiper-pagination-bullet {
-  width: 8px;
-  height: 8px;
-  background: rgba(0, 0, 0, 0.2);
-  opacity: 1;
-  margin: 0 4px;
-}
-
-.ab-intro-pagination .swiper-pagination-bullet-active {
-  background: #666;
-}
-
-/* 인트로 페이지네이션 - 어두운 배경용 (슬라이드 1) */
-.ab-intro-pagination-dark .swiper-pagination-bullet {
-  background: rgba(255, 255, 255, 0.4);
-}
-
-.ab-intro-pagination-dark .swiper-pagination-bullet-active {
-  background: #fff;
-}
