+++ apps/web/src/app/ab-test/_components/steps/index.ts
@@ -1,7 +1,4 @@
 export { default as ABAllergyStep } from './ABAllergyStep';
-export { default as ABIntroSlider } from './A/ABIntroSlider';
-export { default as ABIntroStep } from './A/ABIntroStep';
-export { default as ABuseIntroSlider } from './A/ABuseIntroSlider';
 export { default as ConditionStep } from './ConditionStep';
 export { default as IngredientsStep } from './IngredientsStep';
 export { default as IntroStep } from './IntroStep';
